Add `checkJwt` to Confirm JWT is Valid
Add `checkJwt` to confirm JWT is valid and to do the following:

Verify signature is to call a URL where the keys to the signature are
called JSON Web Key Set (JWKS). This URL is public, and each Auth0
tenant has one. The application would use the JWKS to verify that the
signature in the JWT matches.

Once the signature is verified, there are three standard claims, that
now need to be validated here is how the claims they are going to be
validated and how they are validated:

 1. exp - Expiration and confirm it hasn't expired
 2. iss - Issued by which then needs to confirm it matches your Auth0
          domain
 3. aud - Audience to make sure it matches your clientID