Add `scopes` Entry to `localStorage`
Add the `scopes` entry to `localStorage` so that they can be accessible
throw the application and be able to check if the user can access that
part of the application.

Furthermore, users can not update the `localStorage` locally as the
access token is sent to the server to check if the scopes are allowed
still. The `localStorage` scope is for convenience and not have to
decrypt the JWT for each time we need to check the scope.