Update Auth Class to Storge Session in Memory
Update the Auth Class to storge the session in memory to show how that
works and is more secure than local storage since the session leaves in
memory and is hard to temper. However, there is a drawback where the
session only exists in a single tab, and the session no more once the
user closes the browser.