Add Check Session & Renew Tokens
Add a check if there is a session, and if so, renew tokens before they
expire. This way, the user can stay login.

Some warnings:

 - Silent Auth requires third-party cookies to enable some browsers such
as Safari to have those turn off by default. To Avoid the third-party
cookie issue, you would need to set up a custom domain.

 - Silent Auth does not work with Third-party by default. The developer
needs to go to each identity provider and provide them with a key they
need to create for their application; otherwise, if the user refreshes
or goes to another tab, they not be login anymore.